Split class structure by sequence source
Implemented new multiple inheritance class structure for easy distinction of the several types of workflows
